# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Karigar Service Marketplace API",
        "description": "RESTful API for Karigar - A service marketplace connecting customers with skilled service providers (electricians, plumbers, carpenters, etc.)",
        "contact": {
            "name": "Karigar API Support",
            "email": "support@karigar.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Karigar API Server"
        }
    ],
    "paths": {
...
        "/api/providers": {
            "get": {
                "tags": [
                    "Service Providers"
                ],
                "summary": "Get all service providers",
                "description": "Retrieve list of service providers with optional filters (category, location, search)",
                "operationId": "386f3ddc2b7b26effa6a559287ccbe4c",
                "parameters": [
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter by service category",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Filter by location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by name or service",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Service providers retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/providers/{id}": {
            "get": {
                "tags": [
                    "Service Providers"
                ],
                "summary": "Get provider details",
                "description": "Get detailed information about a specific service provider including services and reviews",
                "operationId": "35b1e0194318551d5d5794e6f71ff577",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Provider ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Provider details retrieved successfully"
                    },
                    "404": {
                        "description": "Provider not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/providers/me": {
            "get": {
                "tags": [
                    "Service Providers"
                ],
                "summary": "Get own provider profile",
                "description": "Service provider can view their own profile and services",
                "operationId": "9f5c96056dda28ed2bab32a79245c8b2",
                "responses": {
                    "200": {
                        "description": "Profile retrieved successfully"
                    },
                    "403": {
                        "description": "Access denied - not a service provider"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Service Providers"
                ],
                "summary": "Update own provider profile",
                "description": "Service provider can update their profile information",
                "operationId": "45031d0eec6ca0228cd7e7c1afed4473",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+1234567890"
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "Electrician"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "New York, NY"
                                    },
                                    "availability": {
                                        "type": "string",
                                        "example": "Mon-Fri 9AM-5PM"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Profile updated successfully"
                    },
                    "403": {
                        "description": "Access denied"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}